change int to float in 2-

+++ 2-Fahrenheit_Celsius.c
@@ -4,8 +4,8 @@
 
 int main ()
 {
-  int fahr, celsius;
-  int lower, upper, step;
+  float fahr, celsius;
+  float lower, upper, step;
 
   lower = 0; /* lower limit of temperature scale */
   upper = 300; /* upper limit */
